Move ATN-related files into own folder
