+++ internal/generator/atn_generator.go
@@ -5,11 +5,12 @@
 package generator
 
 import (
+	"typefox.dev/fastbelt/internal/atn"
 	"typefox.dev/fastbelt/internal/grammar"
 )
 
 func GenerateATN(grammr grammar.Grammar, packageName string, tokenTypes GenerateTokenTypesResult) string {
-	atn, _ := CreateATN(grammr, tokenTypes.TokenTypeIds)
-	source := EmitGoSource(packageName, "BuildATN", "typefox.dev/fastbelt/parser", atn, tokenTypes)
+	a, _ := atn.CreateATN(grammr, tokenTypes.TokenTypeIds)
+	source := atn.EmitGoSource(packageName, "BuildATN", "typefox.dev/fastbelt/parser", a, tokenTypes.TokenTypeVarNames)
 	return FormatIfPossible(source.String())
 }
